requirements.txt
Django==2.1.4
argparse==1.2.1
dj-database-url==0.5.0
djangorestframework==3.8.2
gunicorn==19.9.0
mutagen==1.41.1
pytz==2018.7
whitenoise==4.1.2
wsgiref==0.1.2
